build: Restructure (mesonify) MORT's firmware
