--- src/bsp_atmega328p/external/printf/ATTRIBUTION.md
@@ -1,4 +1,0 @@
-# Attribution: printf
-
-The code in this directory was produced by Marco Paland and sourced from
-[GitHub](https://github.com/mpaland/printf).
