+++ src/hw.c
