+++ src/hw.h
@@ -1,8 +1,8 @@
 #ifndef _CALEBRJC_MORT_HW_H_
 #define _CALEBRJC_MORT_HW_H_
 
-#include "bsp/io/io.h"
-#include "bsp/usart/usart.h"
+#include <bsp/io.h>
+#include <bsp/usart.h>
 
 #define MORT_USART     BSP_USART0
 #define MORT_DEBUG_LED BSP_PB5
