+++ subprojects/printf.wrap
@@ -1,0 +1,7 @@
+[wrap-git]
+url = https://github.com/calebrjc/printf.git
+revision = master
+depth = 1
+
+[provide]
+printf = printf_dep
